refactor: create cutom hook in inbox
