+++ frontend/src/routes/inbox/index.tsx
@@ -1,9 +1,7 @@
 import InboxList from "@/features/inbox/InboxList";
-import { inboxTodosAtom } from "@/stores/todoAtom";
 import { createFileRoute } from "@tanstack/react-router";
-import axios from "axios";
-import { useSetAtom } from "jotai";
-import { useCallback, useEffect } from "react";
+import { useEffect } from "react";
+import { useInboxTodos } from "@/features/inbox/hooks/useInboxTodos";
 
 import SearchBox from "@/components/SearchBox";
 
@@ -11,48 +9,18 @@ export const Route = createFileRoute("/inbox/")({
   component: InboxIndex,
 });
 
-const apiUrl = `${import.meta.env.VITE_BACKEND_API_URL}/inbox`;
-
 function InboxIndex() {
-  const setTodos = useSetAtom(inboxTodosAtom);
-
-  const getTodos = useCallback(async () => {
-    const response = await axios.get(apiUrl);
-    setTodos(response.data);
-  }, [setTodos]);
+  const {
+    getTodos,
+    handleCreate,
+    handleEdit,
+    handleToggleComplete,
+    handleDelete,
+  } = useInboxTodos();
 
   useEffect(() => {
     getTodos();
   }, [getTodos]);
-
-  const handleCreate = async (todo: { name: string; date: string }) => {
-    await axios.post(apiUrl, todo);
-    getTodos();
-  };
-
-  const handleEdit = async (updates: {
-    id: string;
-    name: string;
-    date: string;
-  }) => {
-    await axios.put(`${apiUrl}/${updates.id}`, {
-      name: updates.name,
-      date: updates.date,
-    });
-    getTodos();
-  };
-
-  const handleToggleComplete = async (id: string, completed: boolean) => {
-    await axios.patch(`${apiUrl}/${id}/completed`, {
-      completed,
-    });
-    getTodos();
-  };
-
-  const handleDelete = async (id: string) => {
-    await axios.delete(`${apiUrl}/${id}`);
-    getTodos();
-  };
 
   return (
     <>
